# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"GraphDb API"},"basePath":"/","paths":{"/v1/Relationships/DnsChildren":{"get":{"tags":["DnsChildren"],"operationId":"V1RelationshipsDnsChildrenGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/DnsChildRelationship"}}}}},"post":{"tags":["DnsChildren"],"operationId":"V1RelationshipsDnsChildrenPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"relationships","in":"body","required":false,"schema":{"type":"array","items":{"$ref":"#/definitions/DnsChildRelationship"}}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["DnsChildren"],"operationId":"V1RelationshipsDnsChildrenDelete","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"relationship","in":"body","required":false,"schema":{"$ref":"#/definitions/DnsChildRelationship"}}],"responses":{"200":{"description":"Success"}}}},"/v1/Relationships/DnsPtrs":{"get":{"tags":["DnsPtrs"],"operationId":"V1RelationshipsDnsPtrsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/DnsPtrRelationship"}}}}},"post":{"tags":["DnsPtrs"],"operationId":"V1RelationshipsDnsPtrsPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"relationships","in":"body","required":false,"schema":{"type":"array","items":{"$ref":"#/definitions/DnsPtrRelationship"}}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["DnsPtrs"],"operationId":"V1RelationshipsDnsPtrsDelete","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"relationship","in":"body","required":false,"schema":{"$ref":"#/definitions/DnsPtrRelationship"}}],"responses":{"200":{"description":"Success"}}}},"/v1/Nodes/Domains":{"get":{"tags":["Domains"],"operationId":"V1NodesDomainsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/DomainNode"}}}}},"post":{"tags":["Domains"],"operationId":"V1NodesDomainsPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"nodes","in":"body","required":false,"schema":{"type":"array","items":{"$ref":"#/definitions/DomainNode"}}}],"responses":{"200":{"description":"Success"}}}},"/v1/Nodes/Domains/{name}":{"get":{"tags":["Domains"],"operationId":"V1NodesDomainsByNameGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"name","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/DomainNode"}}}},"delete":{"tags":["Domains"],"operationId":"V1NodesDomainsByNameDelete","consumes":[],"produces":[],"parameters":[{"name":"name","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/v1/Nodes/IpAddresses":{"get":{"tags":["IpAddresses"],"operationId":"V1NodesIpAddressesGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/IpAddressNode"}}}}},"post":{"tags":["IpAddresses"],"operationId":"V1NodesIpAddressesPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"nodes","in":"body","required":false,"schema":{"type":"array","items":{"$ref":"#/definitions/IpAddressNode"}}}],"responses":{"200":{"description":"Success"}}}},"/v1/Nodes/IpAddresses/{ipAddress}":{"get":{"tags":["IpAddresses"],"operationId":"V1NodesIpAddressesByIpAddressGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"ipAddress","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/IpAddressNode"}}}},"delete":{"tags":["IpAddresses"],"operationId":"V1NodesIpAddressesByIpAddressDelete","consumes":[],"produces":[],"parameters":[{"name":"ipAddress","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/v1/Nodes":{"get":{"tags":["Nodes"],"operationId":"V1NodesGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/INode"}}}}}},"/v1/Relationships":{"get":{"tags":["Relationships"],"operationId":"V1RelationshipsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"skip","in":"query","required":false,"type":"integer","format":"int32"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/INodeRelationship"}}}}}},"/v1/Status":{"get":{"tags":["Status"],"operationId":"V1StatusGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}},"head":{"tags":["Status"],"operationId":"V1StatusHead","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}}}}},"definitions":{"DnsChildRelationship":{"type":"object","properties":{"childDomain":{"type":"string"},"parentDomain":{"type":"string"},"type":{"type":"string","readOnly":true}}},"DnsPtrRelationship":{"type":"object","properties":{"ipAddress":{"type":"string"},"domain":{"type":"string"},"type":{"type":"string","readOnly":true}}},"DomainNode":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string","readOnly":true}}},"IpAddressNode":{"type":"object","properties":{"ipAddress":{"type":"string"},"type":{"type":"string","readOnly":true}}},"INode":{"type":"object","properties":{"type":{"type":"string","readOnly":true}}},"INodeRelationship":{"type":"object","properties":{"type":{"type":"string","readOnly":true}}}},"securityDefinitions":{}}